# oas/services/snd/openapi.yaml
openapi: 3.0.3
info:
  title: BThwani — Help (SND) API (سند)
  version: '2025-02-01'
  description: |
    On-demand help service (SND - سند) for the BThwani platform.
    
    This service provides instant help requests and specialized services with pricing, routing, chat, and proof-of-close workflows.
    
    Features:
    - Request creation (instant|specialized)
    - Pricing engine (instant only, ranges/caps)
    - Routing engine (captain/specialized/manual)
    - Encrypted chat with phone masking
    - Proof-of-close with 6-digit codes (instant only)
    - Wallet integration (ledger entries only, no bank payouts)
    - Admin configuration and KPIs
    - Support dispute resolution
    
    Service Code: SRV-SND-01
    Namespace: /api/snd/*
    Timezone: Asia/Aden
    Default Currency: YER
  contact:
    name: BThwani Support
    email: support@bthwani.com
servers:
  - url: https://api-dev.bthwani.com
    description: Development environment
  - url: https://api-stage.bthwani.com
    description: Staging environment
  - url: https://api.bthwani.com
    description: Production environment
...
paths:
...
  /api/snd/captain/requests:
    get:
      operationId: snd_requests_list_captain
      summary: List instant requests for captain
      description: List instant requests available for captain (instant type only)
      tags:
        - APP-CAPTAIN
      x-action-id: list_captain_requests
      x-entity-code: SRV-SND-01
      x-guards: PASS(Auth,Problem,Pagination)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Requests list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SndRequestList'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/captain/requests/{request_id}:
    get:
      operationId: snd_request_get_captain
      summary: Get request details
      description: Get instant request details for captain
      tags:
        - APP-CAPTAIN
      x-action-id: get_request
      x-entity-code: SRV-SND-01
      x-guards: PASS(Auth,Problem,PathParam)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SndRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/captain/requests/{request_id}/accept:
    post:
      operationId: snd_request_accept_captain
      summary: Accept instant request
      description: Accept an instant request (captain only)
      tags:
        - APP-CAPTAIN
      x-action-id: accept_request
      x-entity-code: SRV-SND-01
      x-guards: PASS(Idem,Auth,Problem)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Idempotency-Key'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptRequest'
      responses:
        '200':
          description: Request accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SndRequest'
        '400':
          description: Invalid request type or status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/captain/requests/{request_id}/close-code:
    post:
      operationId: snd_request_generate_close_code
      summary: Generate close code
      description: Generate 6-digit close code for instant request (captain only)
      tags:
        - APP-CAPTAIN
      x-action-id: generate_close_code
      x-entity-code: SRV-SND-01
      x-guards: PASS(Idem,Auth,Problem)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Idempotency-Key'
      responses:
        '200':
          description: Close code generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  close_code:
                    type: string
                    pattern: '^[0-9]{6}$'
                    description: 6-digit close code
                    example: '123456'
        '400':
          description: Invalid request type or already closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/captain/requests/{request_id}/status:
    post:
      operationId: snd_request_status_update_captain
      summary: Update request status
      description: Update request status (captain only)
      tags:
        - APP-CAPTAIN
      x-action-id: update_status
      x-entity-code: SRV-SND-01
      x-guards: PASS(Idem,Auth,Problem)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Idempotency-Key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestStatus'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SndRequest'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/captain/requests/{request_id}/messages:
    get:
      operationId: snd_messages_list_captain
      summary: List chat messages
      description: List chat messages for a request (captain)
      tags:
        - APP-CAPTAIN
      x-action-id: list_messages
      x-entity-code: SRV-SND-01
      x-guards: PASS(Auth,Problem,Pagination)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Chat messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageList'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
    post:
      operationId: snd_message_create_captain
      summary: Send chat message
      description: Send a chat message (captain)
      tags:
        - APP-CAPTAIN
      x-action-id: create_message
      x-entity-code: SRV-SND-01
      x-guards: PASS(Idem,Auth,Problem)
      x-rbac-role: captain
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Idempotency-Key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatMessage'
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/admin/config:
    get:
      operationId: snd_admin_config_get
      summary: Get SND configuration
      description: Get SND service configuration (admin only)
      tags:
        - DASH-ADMIN
      x-action-id: get_config
      x-entity-code: SRV-SND-01
      x-rbac-role: admin
      parameters:
        - $ref: '#/components/parameters/ConfigScope'
        - $ref: '#/components/parameters/ScopeValue'
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
    patch:
      operationId: snd_admin_config_update
      summary: Update SND configuration
      description: Update SND service configuration (admin only, step-up required)
      tags:
        - DASH-ADMIN
      x-action-id: update_config
      x-entity-code: SRV-SND-01
      x-guards: PASS(Idem,Auth,StepUp,Problem)
      x-rbac-role: admin
      parameters:
        - $ref: '#/components/parameters/Idempotency-Key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfig'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/admin/pricing:
    patch:
      operationId: snd_admin_pricing_update
      summary: Update pricing profile
      description: Update pricing profile for instant requests (admin only, step-up required)
      tags:
        - DASH-ADMIN
      x-action-id: update_pricing
      x-entity-code: SRV-SND-01
      x-guards: PASS(Idem,Auth,StepUp,Problem)
      x-rbac-role: admin
      parameters:
        - $ref: '#/components/parameters/Idempotency-Key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePricingProfile'
      responses:
        '200':
          description: Pricing profile updated successfully
          content:
            application/json:
              schema:
                type: object
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
  /api/snd/admin/kpis:
    get:
      operationId: snd_admin_kpis
      summary: Get SND KPIs
      description: Get SND service KPIs and metrics (admin/ops/finance only)
      tags:
        - DASH-ADMIN
      x-action-id: get_kpis
      x-entity-code: SRV-SND-01
      x-rbac-role: admin
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: KPIs and metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SndKPIs'
        default:
          $ref: '#/components/responses/Problem'
      security:
        - bearerAuth: []
...
components:
  schemas:
    SndRequest:
      type: object
      required:
        - id
        - requester_id
        - type
        - title
        - description
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        requester_id:
          type: string
        type:
          type: string
          enum:
            - instant
            - specialized
        category_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          maxLength: 500
        description:
          type: string
        images:
          type: array
          items:
            type: string
          nullable: true
        location:
          type: object
          nullable: true
          properties:
            lat:
              type: number
            lon:
              type: number
        address:
          type: string
          maxLength: 500
          nullable: true
        status:
          type: string
          enum:
            - pending
            - pricing_review
            - routed
            - accepted
            - in_progress
            - completed
            - cancelled
            - escalated
            - disputed
            - closed
        routing_type:
          type: string
          enum:
            - captain
            - specialized_provider
            - manual_queue
          nullable: true
        assigned_captain_id:
          type: string
          nullable: true
        assigned_provider_id:
          type: string
          nullable: true
        price_min_yer:
          type: integer
          description: Minimum price in YER (minor units)
          nullable: true
        price_max_yer:
          type: integer
          description: Maximum price in YER (minor units)
          nullable: true
        price_final_yer:
          type: integer
          description: Final price in YER (minor units)
          nullable: true
        pricing_requires_review:
          type: boolean
          default: false
        close_code:
          type: string
          pattern: '^[0-9]{6}$'
          nullable: true
        close_recipient_name:
          type: string
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        ledger_transaction_id:
          type: string
          format: uuid
          nullable: true
        priced_at:
          type: string
          format: date-time
          nullable: true
        routed_at:
          type: string
          format: date-time
          nullable: true
        accepted_at:
          type: string
          format: date-time
          nullable: true
        in_progress_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        escalation_reason:
          type: string
          nullable: true
        resolution_time_minutes:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateRequestStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - pending
            - pricing_review
            - routed
            - accepted
            - in_progress
            - completed
            - cancelled
            - escalated
            - disputed
            - closed
        reason:
          type: string
          maxLength: 500
...
    AcceptRequest:
      type: object
      properties:
        notes:
          type: string
          maxLength: 500
    SndRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SndRequest'
        next_cursor:
          type: string
          nullable: true
    ChatMessage:
      type: object
      required:
        - id
        - request_id
        - sender_id
        - recipient_id
        - direction
        - body_encrypted
        - created_at
      properties:
        id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        sender_id:
          type: string
        recipient_id:
          type: string
        direction:
          type: string
          enum:
            - requester_to_captain
            - captain_to_requester
            - requester_to_provider
            - provider_to_requester
        body_encrypted:
          type: string
          description: Decrypted message body (encrypted at rest)
        phone_masked:
          type: string
          nullable: true
        links_masked:
          type: array
          items:
            type: string
          nullable: true
        is_read:
          type: boolean
          default: false
        read_at:
          type: string
          format: date-time
          nullable: true
        is_urgent:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
    CreateChatMessage:
      type: object
      required:
        - body
      properties:
        body:
          type: string
          maxLength: 2000
        is_urgent:
          type: boolean
          default: false
    ChatMessageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        next_cursor:
          type: string
          nullable: true
    UpdateConfig:
      type: object
      required:
        - scope
        - key
        - value
      properties:
        scope:
          type: string
          enum:
            - global
            - region
            - category
            - category_region
        scope_value:
          type: string
        category_id:
          type: string
          format: uuid
        key:
          type: string
        value:
          type: object
        description:
          type: string
        is_active:
          type: boolean
          default: true
    UpdatePricingProfile:
      type: object
      required:
        - scope
        - min_price_yer
        - max_price_yer
      properties:
        scope:
          type: string
          enum:
            - global
            - region
            - category
            - category_region
        scope_value:
          type: string
        category_id:
          type: string
          format: uuid
        min_price_yer:
          type: integer
          minimum: 0
          description: Minimum price in YER (minor units)
        max_price_yer:
          type: integer
          minimum: 0
          description: Maximum price in YER (minor units)
        requires_review:
          type: boolean
          default: false
        is_active:
          type: boolean
          default: true
...
    SndKPIs:
      type: object
      properties:
        requests_total:
          type: integer
        requests_instant_count:
          type: integer
        requests_specialized_count:
          type: integer
        requests_instant_vs_specialized_ratio:
          type: number
        sla_compliance_rate:
          type: number
        cancellation_rate:
          type: number
        dispute_rate:
          type: number
        reroute_count:
          type: integer
        avg_resolution_time_minutes:
          type: number
        requests_by_status:
          type: object
          additionalProperties:
            type: integer
        requests_by_type:
          type: object
          additionalProperties:
            type: integer
    Problem:
      type: object
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
  parameters:
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Status:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum:
          - pending
          - pricing_review
          - routed
          - accepted
          - in_progress
          - completed
          - cancelled
          - escalated
          - disputed
          - closed
...
    CategoryId:
      name: category_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Idempotency-Key:
      name: Idempotency-Key
      in: header
      required: true
      description: Idempotency key (UUID v4) for mutating operations
      schema:
        type: string
        format: uuid
    ConfigScope:
      name: scope
      in: query
      required: false
      schema:
        type: string
        enum:
          - global
          - region
          - category
          - category_region
    ScopeValue:
      name: scope_value
      in: query
      required: false
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Problem:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT